+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - feat: Respect existing indentation when updating project `package.json`
 - feat: Handle typescript 5 in peerDependencies
 - feat: Relax auto-fixable `quotes` and `curly` rules to "warn"
+- fix: `pkg.hxmstyle.options` should not be auto-removed — as auto-detction
+  often sees false-negatives in monorepos
 
 ## 7.0.0